Don't enforce xvfb for generating the man pages
Summary:
D7677 made it mandatory to use `xvfb` for generating the man pages. This
causes a few issues:
 - It enforces installing `xvfb` which comes with a lot of dependencies, even if `bitcoin-qt` is not built.
 - `xvfb-run` is not designed to run parallel, and will fail on some systems due to race when running `ninja install-manpages` (even with the `-a` flag).

This diff removes the mandatory `xvfb` and adds a README to instruct the
user that it might be required and how to install it.

Test Plan:
Read and try the instructions.

  ./contrib/teamcity/build-configurations.py build-docs

On Archlinux:
  ninja install-manpages
Should no longer fail.

Reviewers: #bitcoin_abc, jasonbcox

Reviewed By: #bitcoin_abc, jasonbcox

Differential Revision: https://reviews.bitcoinabc.org/D7913