[CI] Add a configuration for running on native OSX agents
Summary:
This configuration runs the unit tests, functional tests (non extended)
and secp256k1 tests. The timeout is kept high deliberately because the
agent is slow. For now it is not really osx specific, except for the
timeout being very high.

Test Plan: Run the build on an OSX agent.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7951