[land-bot] Add a sanity rebase after running smoke tests
Summary:
This ensures that any non-colliding changes pushed upstream within the few minute
window that smoke tests were running will not break the `git push`. While we do make
the assumption that tests still pass after this, breaking changes from automated
diffs should be exceedingly rare.  For human-authored changes, we have a land queue to
help mitigate this.

This is copied from `automated-commits.sh` to bring this script up to
par with the sanity checking we already do there. The next step is to cleanup
`automated-commits.sh` so that it calls `land-patch.sh` directly.

Depends on D7467 and D7911

Test Plan:
  cmake -GNinja -DENABLE_SOURCE_CONTROL_TOOLS_TESTS=ON ..
  ninja check-source-control-tools

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7468