[CI] Fail fast the builds that print error to stdout
Summary:
Some builds print errors to stdout. As a consequence we can't jump a some specific test failure when an error is thrown, and the error message displayed to the user is the build tail.
This diff makes these build stop on first error, so that the build tail will contain the actual issue and the user won't need to look at the full teamcity build log.

Test Plan:
Edit 2 files to remove some braces, then:
  ./contrib/teamcity/build-configurations.py build-clang-tidy
Check the build stops after the first missing brace.

  ./contrib/teamcity/build-configurations.py build-clang-asan
  ./contrib/teamcity/build-configurations.py build-clang-ubsan
  ./contrib/teamcity/build-configurations.py build-clang-tsan

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7973