rcp: update validateaddress RPCExamples to CashAddr
Summary:
This is a backport of a PR that factors out an address used as an example in the help of RPC commands, and uses a CashAddr instead of a legacy address. The original PR uses an invalid address, but IMO it is more useful to have the documentation of a command provide an example that works, so I used the correct address just converted from legacy to CashAddr.

Backport of Core [[https://github.com/bitcoin/bitcoin/pull/18122 | PR18122]]

Test Plan:
`ninja && ninja	check`

```
bitcoin-cli help getaddressinfo
bitcoin-cli help validateaddress
```

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D7972