qt: Change default size of intro frame
Summary:
Because of the new pruning feature in the intro dialog, the size of the intro frame was too small.
Adjust the size of the dialog after setting the content of all widget.

This is a backport of Core [[https://github.com/bitcoin/bitcoin/pull/16971 | PR16971]]

Test Plan: Run `src/qt/bitcoin-qt -choosedatadir` to show the intro dialog on startup.

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8078