[BUILDBOT] Use tokens as an immediate feedback and anti DoS mechanism
Summary:
Tokens are awarded on a per-user and per-object basis, and are
persistent, which make them a good candidate for an anti-DoS as the bot
can find the previous state and decide based on this.
The current filter allow any approved, verified, and activated non ABC
user to request builds from the bot, with a total limit of 4 requests
per revision. ABC members bypass this rule.

This token mechanism also provide a discrete immediate positive feedback
to the user, improving the UX.

Depends on D8086.

Test Plan:
  pytest

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8087