remove redundant strprintf and add missing sentence in doc
Summary:
This includes a sentence in the documentation of `-stdin` that was missed while backporting [[https://github.com/bitcoin/bitcoin/pull/10997 | PR10997]] (D1042),
and the removal of a redundant `strprintf`. The double space was already removed in D3058.

This is a backport of Core [[https://github.com/bitcoin/bitcoin/pull/13905 | PR13905]]

Test Plan:
`ninja && ninja check`
`bitcoin-cli -help`

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8105