[CMAKE] Link the fuzzer test_runner.py when building a fuzz target
Summary:
The script is linked to the build directory when building `all`, but not
when building `bitcoin-fuzzers` or any other fuzz target. This diffs
builds a list of the fuzz targets and make them depend on the link so
that it is created as needed.

Test Plan:
  cmake -GNinja .. \
    -DCMAKE_C_COMPILER=clang \
    -DCMAKE_CXX_COMPILER=clang++ \
    -DENABLE_SANITIZERS=fuzzer
  ninja bitcoin-fuzzers # Or any fuzz-* target
Check the `test/fuzz/test_runner.py` script is linked to the build tree.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8137