Show addresses for "SendToSelf" transactions
Summary:
With this change, the transactions tab of bitcoin-qt now shows the address for "Payment to yourself" transactions.

This is a backport of Core [[https://github.com/bitcoin/bitcoin/pull/15098 | PR15098]]

Test Plan:
`ninja && src/qt/bitcoin-qt`
Create a receive address on the Receive tab and send some BCH to it on the Send tab.
Verify that the Transactions tab now show the involved addresses.

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, deadalnix

Subscribers: deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8122