Revert "Revert "fuzz: Add process_messages harness""
Summary:
This reverts commit rABC22c85ce64493710b114ad598a594f9a9b25422c5.

This is a backport of core [[https://github.com/bitcoin/bitcoin/pull/18521 | PR18521]]. It is similar to D8376 rebased
on top of D8525 which fixes the tsan race.

Depends on D8634 for the test plan.

Test Plan:
  ninja bitcoin-fuzzers
  ./src/test/fuzz/process_messages <path_to_corpus>

  ninja all check-all

With TSAN:
  ninja check-bitcoin-radix_tests check-bitcoin-coinselector_tests
This will likely trigger the TSAN issue on D8376 but not on this diff.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8635