CConnman: initialise at declaration rather than in Start()
Summary:
Ensure `nMaxOutboundTotalBytesSentInCycle` and `nMaxOutboundCycleStartTime` are initialized even if `CConnman::Start()` is not called.  Prevents failures in test/fuzz/connman when run under valgrind.

`nTotalBytesRecv` and `nTotalBytesSent` are already initialized in the class definition, and `CConnman::Start` is only called once in `AppInitMain`, so removing the reinitialization in `Start` has no impact.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/20408 | core#20408]]

Test Plan:
With TSAN:
`ninja && ninja check check-functional`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10740