[buildbot] Introduce selecting diff builds based on filename regex
Summary:
As this codebase turns more into a mono-repo, we need a way to reduce
CI load since the number of patches is expected to grow by an order of magnitude.

A regex that matches changed filenames is a quick and easy way to determine which tests
may be valuable to run. While this isn't perfect, we can get a huge payoff almost
immediately. Consider this recent example where D8441 ran a bunch of node-related tests
when it clearly is not related in any way. If we expect 3+ devs to work on website(s)
then we cannot tie up the CI running node tests when there is no behavior change.

In addition, this allows us to be more aggressive in running wider-coverage of tests
for specific types of changes. This patch demonstrates that by adding buildbot as a
diff-enabled build. Before this patch, running the buildbot tests would have been a
waste of CI resources as it would run on every revision. After this patch, that is no
longer the case.

Updates to other build configurations will come in future patches so they can be reviewed
more carefully on their own.

Depends on D8473

Test Plan:
  pytest

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D8471