[buildbot] Add revision ID as arg to /buildDiff
Summary:
This provides an easier way for infra devs to identify which revision
may be causing infra outages. It also enables making Conduit calls based on
revision ID in `/buildDiff`, which is the first step to being able to configure
running tests on diffs only when certain files have been changed. This will come
in a followup diff.

Test Plan:
  pytest

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D8459