Only run node-related builds on diffs where the node software has changed
Summary:
This ensures changes to the docs, websites, and other infra do not
kick off a bunch of tests that only pertain to the node software. While these
heuristics may not be perfect, any breaking change will still be caught by
CI after changes are landed.

Test Plan:
There is currently no good way to test this, so just read it
carefully for now.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D8483