[buildbot] Fix getcommitpaths result type
Summary:
Despite the Conduit testing dashboard returning a dictionary, actually
calling the endpoint directly with curl gives a list. This may have something to
do with the way that Conduit is written in PHP and printing the list out with
integer indexing, which is the default in PHP.

This fixes builds not triggering correctly on revisions.

Test Plan: `pytest`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D8482