qt: unlock wallet "OK" button bugfix
Summary:
When trying to send a transaction from an encrypted wallet, the ask passphrase dialog would not allow the user to click the "OK" button and proceed. Therefore it was impossible to send a transaction through the gui. It was not enabling the "OK" button after the passphrase was entered by the user, because it was using the same form validation logic as the "Change passphrase" flow.

This bug was introduced in D10737 (v0.24.7).

This is a backport of [[https://github.com/bitcoin-core/gui/pull/138 | core-gui#138]]

Test Plan: In `bitcoin-qt`, try to send a transaction using an encrypted wallet. Verify that the "OK" button is not disabled in the passphrase dialog after typing at least one character in the passphrase field.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D10831