[avalanche] Fix race condition during destruction in poll_and_response test
Summary: Previously, the Processor would be destroyed at the end of the test execution, but the callback thread would still be running, possibly calling into a destroyed Processor. To solve this problem, we move the Processor ownership to the test framework.

Test Plan:
  ninja all check-avalanche

Run it multiple time to ensure the race is gone.

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D8603