[avalanche] Initiate avalanche when the service bit is present and we have a stake.
Summary:
When a new node connects itself and supports avalanche, then we send an avahello message. This message simply contains a delegation and a signature proving we own the final key.

Depends on D8861

In order to ensure the signature is secure, we need extra entropy from our peer, which is exactly what D8861 does.

Test Plan:
  ninja all check-all

Extended the avalanche test to check thta avahello messages are sent properly.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D8862