test: run mempool_expiry.py even with wallet disabled
Summary:
```
Test coverage is also extended in this commit to:
Ensure that another transaction in the mempool is not evicted when other transactions expire. Note: this other transaction does not have any ancestors in the mempool that expired. Otherwise it would be evicted from the mempool, and we already test this case.
```

Backport of [[https://github.com/bitcoin/bitcoin/pull/20276 | core#20276]].

Test Plan:
  ./test/functional/test_runner.py mempool_expiry

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D12423