doc: Convert depends options list from html to markdown
Summary:
This makes it easier to read in `less`, which is important for install
instructions.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/20681 | core#20681]]

Test Plan: proofreading. Check that all the options are still there and no new option has been added while solving the cherry-pick conflict.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D12090