gui: Shortcut to close ModalOverlay
Summary:
Set a shortcut to close the modal overlay in bitcoin-qt  with the Esc key.
The rationale is that it makes testing easier, when you need to start the application multiple times in a row.

This is a backport of Core [[https://github.com/bitcoin/bitcoin/pull/17998 | PR17998]]

Test Plan: `ninja && src/qt/bitcoin-qt`

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D8689