[BUILDBOT] Temporarly fix the 502 error with embedded shields.io link
Summary:
The status panel gets a 502 when requesting the badge from shields.io.
This is unclear where the issue resides because the error only occurs
with phabricator (the link works in a browser and curl gets no error).

This diff just use the simpleicons.org Teamcity logo instead of our
custom one, which works around the issue.

Test Plan:
  pytest

Reviewers: #bitcoin_abc, majcosta

Reviewed By: #bitcoin_abc, majcosta

Differential Revision: https://reviews.bitcoinabc.org/D8759